refactor(term): move terminal related logic behind into term module
